+++ stdnet/__init__.py
@@ -10,11 +10,13 @@ __author__ = "Luca Sbardella"
 __contact__ = "luca.sbardella@gmail.com"
 __homepage__ = "https://github.com/lsbardel/python-stdnet"
 CLASSIFIERS = [
-               'Development Status :: 4 - Beta',
+               'Development Status :: 5 - Production/Stable',
                'Environment :: Plugins',
+               'Environment :: Web Environment',
                'Intended Audience :: Developers',
                'License :: OSI Approved :: BSD License',
                'Operating System :: OS Independent',
+               'Programming Language :: Python',
                'Programming Language :: Python :: 2.6',
                'Programming Language :: Python :: 2.7',
                'Programming Language :: Python :: 3.1',